+++ FleetManager/fleetmanager.client/src/App.css
@@ -1,11 +1,37 @@
 ﻿#root {
-    max-width: 1280px;
+    width: 100%;
     margin: 0 auto;
-    padding: 2rem;
     text-align: center;
+    height: 100vh;
 }
 
 th, td {
     padding-left: 1rem;
     padding-right: 1rem;
+}
+table {
+    margin: 0 auto;
+}
+nav {
+    margin-bottom: 2rem;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    
+    
+    padding: 1rem;
+    background-color: #f8f9fa;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    font-family: Arial, sans-serif;
+
+    text-transform: uppercase;
+    font-size: 1.2rem;
+    
+       
+}
+#navLinks {
+    min-width: 50%;
+    display: flex;
+    align-items: center;
+    justify-content: space-around;
 }
