+++ FleetManager/package.json
@@ -1,0 +1,11 @@
+{
+  "dependencies": {
+    "react-router": "^7.5.3",
+    "react-router-dom": "^7.5.3"
+  },
+  "devDependencies": {
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
+    "@types/react-router-dom": "^5.3.3"
+  }
+}
